LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/textfieldselection/selection-not-application-expected.txt


FAIL text field selection for the input hidden Type error
FAIL text field selection for the input email assert_equals: expected (object) null but got (number) 0
FAIL text field selection for the input datetime-local assert_equals: expected (object) null but got (number) 0
FAIL text field selection for the input date assert_equals: expected (object) null but got (number) 0
FAIL text field selection for the input month assert_equals: expected (object) null but got (number) 0
FAIL text field selection for the input week assert_equals: expected (object) null but got (number) 0
FAIL text field selection for the input time assert_equals: expected (object) null but got (number) 0
FAIL text field selection for the input number assert_equals: expected (object) null but got (number) 0
FAIL text field selection for the input range Type error
FAIL text field selection for the input color Type error
FAIL text field selection for the input checkbox Type error
FAIL text field selection for the input radio Type error
FAIL text field selection for the input file Type error
FAIL text field selection for the input submit Type error
FAIL text field selection for the input image Type error
FAIL text field selection for the input reset Type error
FAIL text field selection for the input button Type error
PASS text field selection for the input text 
PASS text field selection for the input search 
PASS text field selection for the input tel 
PASS text field selection for the input url 
PASS text field selection for the input password 
